+++ src/resources/shaders/raytrace.glsl
@@ -1,10 +1,3 @@
-#pragma once
-#include <string>
-
-namespace HaloSim
-{
-
-const std::string computeShaderSource = R""(
 #version 440 core
 
 #define DISTRIBUTION_UNIFORM 0
@@ -50,6 +43,8 @@ uniform struct camera_t
     int projection;
     int hideSubHorizon;
 } camera;
+
+uniform float multipleScatteringProbability;
 
 const float PI = 3.1415926535;
 
@@ -470,12 +465,28 @@ void main(void)
 
     if (length(resultRay) < 0.0001) return;
 
-    resultRay = -rotationMatrix * resultRay;
+    resultRay = rotationMatrix * resultRay;
+
+    if (multipleScatteringProbability != 0.0 && multipleScatteringProbability > rand())
+    {
+        // Rotation matrix to orient ray/crystal
+        rotationMatrix = getRotationMatrix();
+
+        /* The inverse rotation matrix must be applied because we are
+        rotating the incoming ray and not the crystal itself. */
+        rotatedRayDirection = resultRay * rotationMatrix;
+
+        resultRay = castRayThroughCrystal(rotatedRayDirection, wavelength);
+
+        if (length(resultRay) < 0.0001) return;
+
+        resultRay = rotationMatrix * resultRay;
+    }
 
     // Hide subhorizon rays
-    if (camera.hideSubHorizon == 1 && resultRay.y < 0.0) return;
+    if (camera.hideSubHorizon == 1 && resultRay.y > 0.0) return;
 
-    resultRay = getCameraOrientationMatrix() * resultRay;
+    resultRay = -getCameraOrientationMatrix() * resultRay;
 
     ivec2 resolution = imageSize(outputImage);
     float aspectRatio = float(resolution.y) / float(resolution.x);
@@ -514,6 +525,4 @@ void main(void)
     ivec2 pixelCoordinates = ivec2(resolution.x * normalizedCoordinates.x, resolution.y * normalizedCoordinates.y);
     vec3 cieXYZ = daylightEstimate(wavelength) * vec3(xFit_1931(wavelength), yFit_1931(wavelength), zFit_1931(wavelength));
     storePixel(pixelCoordinates, cieXYZ);
-}
-)"";
 }