+++ src/simulation/simulationEngine.cpp
@@ -7,7 +7,6 @@
 #include "camera.h"
 #include "lightSource.h"
 #include "crystalPopulation.h"
-#include "shaders/raytrace.glsl"
 
 namespace HaloSim
 {
@@ -27,6 +26,7 @@ SimulationEngine::SimulationEngine(
       mCamera(Camera::createDefaultCamera()),
       mLight(LightSource::createDefaultLightSource()),
       mCameraLockedToLightSource(false),
+      mMultipleScatteringProbability(0.0),
       mCrystalRepository(crystalRepository)
 {
 }
@@ -136,6 +136,8 @@ void SimulationEngine::step()
         mSimulationShader->setUniformValue("camera.projection", mCamera.projection);
         mSimulationShader->setUniformValue("camera.hideSubHorizon", mCamera.hideSubHorizon ? 1 : 0);
 
+        mSimulationShader->setUniformValue("multipleScatteringProbability", mMultipleScatteringProbability);
+
         glDispatchCompute(numRays, 1, 1);
     }
 }
@@ -176,7 +178,7 @@ void SimulationEngine::initialize()
 void SimulationEngine::initializeShader()
 {
     mSimulationShader = std::make_unique<QOpenGLShaderProgram>();
-    mSimulationShader->addCacheableShaderFromSourceCode(QOpenGLShader::ShaderTypeBit::Compute, computeShaderSource.c_str());
+    mSimulationShader->addCacheableShaderFromSourceFile(QOpenGLShader::ShaderTypeBit::Compute, ":/shaders/raytrace.glsl");
     if (mSimulationShader->link() == false)
     {
         throw std::runtime_error(mSimulationShader->log().toUtf8());
@@ -212,6 +214,17 @@ void SimulationEngine::pointCameraToLightSource()
     clear();
     mCamera.yaw = 0.0f;
     mCamera.pitch = mLight.altitude;
+}
+
+void SimulationEngine::setMultipleScatteringProbability(double probability)
+{
+    clear();
+    mMultipleScatteringProbability = static_cast<float>(std::min(std::max(probability, 0.0), 1.0));
+}
+
+double SimulationEngine::getMultipleScatteringProbability() const
+{
+    return static_cast<double>(mMultipleScatteringProbability);
 }
 
 } // namespace HaloSim