+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2.3.0 - 2019-07-13
+
+### Added
+- Added menu bar with option to save simulation result as image
+- Added slider for setting the probability of double scattering
+
+### Changed
+- Replaced Add Population and Remove Population texts with icons
+
 ## 2.2.0 - 2019-07-09
 
 ### Added
@@ -12,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added possibility to add preset crystal populations
 
 ### Changed
-- Disabled "Add population" button when there is only one population in simulation
+- Disabled "Remove population" button when there is only one population in simulation
 - Added multiple crystal populations by default
 
 ## 2.1.1 - 2019-07-08